Data input form updated

+++ kumulus/grails-app/controllers/com/kumulus/controllers/workflow/BackOfficeController.groovy
@@ -50,43 +50,45 @@ class BackOfficeController {
         // NEED TO MOVE ALL THIS TO A SERVICE, CURRENTLY IT IS NOT ATOMIC
         def response = [done: false]
         def data = request.JSON
-        def task = Task.findById(data?.form.taskId)
-        def currency = Currency.findById(data?.form.currency)
-        def documentType = DocumentType.findById(data?.form.documentType)
+        if(data?.form && data.form?.taskId && data.form?.currency && data.form?.documentType) {
+            def task = Task.findById(data.form.taskId)
+            def currency = Currency.findById(data.form.currency)
+            def documentType = DocumentType.findById(data.form.documentType)
 
-        if(task && !task.completed && currency && documentType && data?.form.company && data?.form.identifier) {
-            
-            // update the document
-            Date date
-            if(data.form.date) date = new Date().parse("dd/MM/yyyy", data.form.date) else date = null
-            def document = structureService.updateDocument(task.document, data.form.company, documentType, date, data.form.identifier)
-            
-            // update the line items submitted
-            def updatedLineItems = []
-            data.form?.lineItems.each {
-                date = null
-                if(it?.lineItemDate) date = new Date().parse("dd/MM/yyyy", it.lineItemDate)
-                def lineItem = structureService.updateLineItem(it?.lineItemId, it?.pageId, currency, date, it?.description, it?.quantity, it?.price, it?.amount)
-                updatedLineItems.add(lineItem)
-            }
-            
-            // remove any deleted line items from the database
-            def lineItemsToRemove = document.pages.lineItems.toList().flatten()
-            lineItemsToRemove.removeAll(updatedLineItems)
-            lineItemsToRemove.each { lineItem ->
-                lineItem.page.removeFromLineItems(lineItem)
-                lineItem.delete()
-            }
-            
-            // close the task if requested
-            if(data?.completeTask) {
-                workflowService.completeTask(task)
-                if(task.type==Task.TYPE_PROCESS) {
-                    def newTask = workflowService.createTask(document, Task.TYPE_VALIDATE, permissionsService.getUsername())
-                    workflowService.assignTask(newTask, task.document.project.ownerId)
+            if(task && !task.completed && currency && documentType && data?.form.company && data?.form.identifier) {
+
+                // update the document
+                Date date
+                if(data.form.date) date = new Date().parse("dd/MM/yyyy", data.form.date) else date = null
+                def document = structureService.updateDocument(task.document, data.form.company, documentType, date, data.form.identifier)
+
+                // update the line items submitted
+                def updatedLineItems = []
+                data.form?.lineItems.each {
+                    date = null
+                    if(it?.lineItemDate) date = new Date().parse("dd/MM/yyyy", it.lineItemDate)
+                    def lineItem = structureService.updateLineItem(it?.lineItemId, it?.pageId, currency, date, it?.description, it?.quantity, it?.price, it?.amount)
+                    updatedLineItems.add(lineItem)
                 }
+
+                // remove any deleted line items from the database
+                def lineItemsToRemove = document.pages.lineItems.toList().flatten()
+                lineItemsToRemove.removeAll(updatedLineItems)
+                lineItemsToRemove.each { lineItem ->
+                    lineItem.page.removeFromLineItems(lineItem)
+                    lineItem.delete()
+                }
+
+                // close the task if requested
+                if(data?.completeTask) {
+                    workflowService.completeTask(task)
+                    if(task.type==Task.TYPE_PROCESS) {
+                        def newTask = workflowService.createTask(document, Task.TYPE_VALIDATE, permissionsService.getUsername())
+                        workflowService.assignTask(newTask, task.document.project.ownerId)
+                    }
+                }
+                response.done = true
             }
-            response.done = true
         }
         render response as JSON
     }
